show more component & repository form component add #13
